Refactoring Instructions and Variables
