+++ dockerfile_ast/utils.py
@@ -1,8 +1,33 @@
+from abc import ABCMeta
 import logging
 import sys
 
 
+class DockerfileASTNode(metaclass=ABCMeta):
+    """
+    A node of all possible syntax for Dockerfile AST.
+    """
+    pass
+
+
 def init_logger(stream_level: int, log_filename: str, file_level: int) -> logging.Logger:
+    """
+    Initialize `logging.Logger` for Dockerfile AST.
+
+    Parameters
+    ----------
+    stream_level : int
+        Logging level of StreamHandler.
+    log_filename : str
+        Log file name.
+    file_level : int
+        Logging level of FileHandler.
+
+    Returns
+    -------
+    logging_logger : logging.Logger
+        logging.Logger for Dockerfile AST.
+    """
     logging_logger = logging.getLogger(__name__)
     # ログで出力するレベルを指定
     logging_logger.setLevel(logging.DEBUG)
